refactor: split out client ready callback
